// src/api/nexuscore/model/questCompletionBucket.ts
/**
 * Generated by orval v8.8.0 🍺
 * Do not edit manually.
 * Nexuscore
 * The backend for all Everthorn internal services
 * OpenAPI spec version: 1.0.0
 */

export interface QuestCompletionBucket {
  /** Start of the time bucket in seconds */
  bucket_start_seconds: number;
  /** End of the time bucket in seconds */
  bucket_end_seconds: number;
  /** Number of completions that fall within this bucket */
  count: number;
}